form validation - trip cannot begin in the past check implemented
trip model - type_of_travel field maximum length increased from 15 to 20 to fix bug where one of the choices is longer than 15 characters

templates - Cancel button added to all form views in the traveler app. submit button on forms and edit button on details of various templates in the traveler app made smaller and class changed to standardize the onhover behavior with the rest of the application.